+++ board-back/src/main/java/com/zoo/boardback/domain/comment/application/CommentService.java
@@ -6,10 +6,10 @@ import static com.zoo.boardback.global.error.ErrorCode.COMMENT_NOT_FOUND;
 import com.zoo.boardback.domain.board.dao.BoardRepository;
 import com.zoo.boardback.domain.board.entity.Board;
 import com.zoo.boardback.domain.comment.dao.CommentRepository;
+import com.zoo.boardback.domain.comment.dto.query.CommentQueryDto;
 import com.zoo.boardback.domain.comment.dto.request.CommentCreateRequestDto;
 import com.zoo.boardback.domain.comment.dto.request.CommentUpdateRequestDto;
 import com.zoo.boardback.domain.comment.dto.response.CommentListResponseDto;
-import com.zoo.boardback.domain.comment.dto.response.CommentResponse;
 import com.zoo.boardback.domain.comment.entity.Comment;
 import com.zoo.boardback.domain.user.entity.User;
 import com.zoo.boardback.global.error.BusinessException;
@@ -38,7 +38,7 @@ public class CommentService {
   public CommentListResponseDto getComments(int boardNumber) {
     Board board = boardRepository.findById(boardNumber).orElseThrow(
         () -> new BusinessException(boardNumber, "boardNumber", BOARD_NOT_FOUND));
-    List<CommentResponse> comments = commentRepository.getCommentsList(board);
+    List<CommentQueryDto> comments = commentRepository.getCommentsList(board);
     return CommentListResponseDto.from(comments);
   }
