+++ board-back/src/main/java/com/zoo/boardback/domain/comment/dao/CommentRepository.java
@@ -1,7 +1,7 @@
 package com.zoo.boardback.domain.comment.dao;
 
 import com.zoo.boardback.domain.board.entity.Board;
-import com.zoo.boardback.domain.comment.dto.response.CommentResponse;
+import com.zoo.boardback.domain.comment.dto.query.CommentQueryDto;
 import com.zoo.boardback.domain.comment.entity.Comment;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,10 +10,11 @@ import org.springframework.data.repository.query.Param;
 
 public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
-  @Query("SELECT new com.zoo.boardback.domain.comment.dto.response.CommentResponse(A.commentNumber, B.nickname, B.profileImage, A.content) " +
+  @Query("SELECT " +
+      "new com.zoo.boardback.domain.comment.dto.query.CommentQueryDto(A.commentNumber, B.nickname, B.profileImage, A.content, A.createdAt, A.updatedAt) " +
       "FROM Comment A " +
       "INNER JOIN A.user B " +
       "WHERE A.board = :board " +
       "ORDER BY A.createdAt DESC")
-  List<CommentResponse> getCommentsList(@Param("board") Board board);
+  List<CommentQueryDto> getCommentsList(@Param("board") Board board);
 }
