+++ board-back/src/main/java/com/zoo/boardback/domain/comment/dto/response/CommentListResponseDto.java
@@ -1,5 +1,7 @@
 package com.zoo.boardback.domain.comment.dto.response;
 
+import com.zoo.boardback.domain.comment.dto.query.CommentQueryDto;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.Getter;
@@ -13,13 +15,17 @@ public class CommentListResponseDto {
     this.commentListResponse = commentListResponse;
   }
 
-  public static CommentListResponseDto from(List<CommentResponse> comments) {
+  public static CommentListResponseDto from(List<CommentQueryDto> comments) {
+    DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+
     List<CommentResponse> commentListResponse = comments.stream()
         .map(comment -> CommentResponse.builder()
             .commentNumber(comment.getCommentNumber())
             .nickname(comment.getNickname())
             .profileImage(comment.getProfileImage())
             .content(comment.getContent())
+            .createdAt(comment.getCreatedAt().format(formatter))
+            .updatedAt(comment.getUpdatedAt().format(formatter))
             .build())
         .collect(Collectors.toList());
     return new CommentListResponseDto(commentListResponse);
